tests: what a WASM closure's code object carries, and one file per subject
The two new tests build a closure through `rt_make_closure` — a hand-written
module whose entry is one `MakeClosure` — and read the four fields the site
drops back off its code object: both release tables, the location map and the
merge set. Both fail today, and fail on the assertion:

  left: []  right: [3, 7]

Nothing else moves. A closure built from a blueprint missing those fields
carries real bytecode and returns right answers, which is why the emitter and
the JIT each lost them once before.

src/wasm/tests.rs was 933 lines over ten subjects, so it becomes a directory of
one file per subject, each carrying its call-out and stamp. The two spawn tests
join the new ones in `tests/closure.rs`: they ask the same question of the same
code object, from the worker's side.

Auditing the files that named the old path pulled three more in. emit.rs was 569
lines and sheds its ABI constants to `emit/abi.rs`, the numbers the module and
the host agree on. host.rs's header said `call_primitive` was the main host
function; nothing has emitted a call to it in a long time, and `rt_call` is
what a compiled call reaches. src/wasm/AGENTS.md's file table had drifted from
the directory, and its known-issues list carried an unreproducible one.
